backend/requirements.txt
argh==0.26.2
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.5
django-dramatiq==0.9.1
dramatiq==1.8.1
future==0.18.2
gevent==20.5.0
greenlet==0.4.15
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
multidict==4.7.5
oauthlib==3.1.0
pathtools==0.1.2
pika==1.1.0
prometheus-client==0.7.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dotenv==0.13.0
PyTumblr==0.1.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
watchdog==0.8.3
watchdog-gevent==0.1.0
yarl==1.4.2
